project/requirements.txt
djangorestframework==3.16.1
langchain==0.3.27
neo4j==6.0.2
openai==0.28.0
openpyxl==3.1.5
pandas==2.3.3
pandas-stubs==2.3.2.250926
pip==25.2
setuptools==65.5.0
